% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-utility-functions.R
\name{annotateMz,numeric,DataFrameOrEquivalent-method}
\alias{annotateMz,numeric,DataFrameOrEquivalent-method}
\alias{annotateMz}
\alias{annotateMz,DataFrameOrEquivalent,DataFrameOrEquivalent-method}
\alias{annotateMz,numericOrDataFrameOrEquivalent,CompDb-method}
\title{Annotate m/z values for ion adducts}
\usage{
\S4method{annotateMz}{numeric,DataFrameOrEquivalent}(object, compounds, adduct = "[M+H]+", ppm = 10, ...)

\S4method{annotateMz}{DataFrameOrEquivalent,DataFrameOrEquivalent}(object, compounds, adduct = "[M+H]+", ppm = 10, mzcol = "mz", ...)

\S4method{annotateMz}{numericOrDataFrameOrEquivalent,CompDb}(object, compounds, adduct = "[M+H]+", ppm = 10, ...)
}
\arguments{
\item{object}{either a \code{numeric} of m/z values or a \code{data.frame} (or
equivalent) with a column containing the m/z values.}

\item{compounds}{either a \code{data.frame} (or equivalent) or a \code{\link[=CompDb]{CompDb()}} with
the reference annotations. A column named \code{"exactmass"} is mandatory if
\code{compounds} is a \code{data.frame}.}

\item{adduct}{adduct definition. A \code{character} defining the adduct(s) to
look for or a \code{data.frame} with adduct definition. See help on
\code{\link[=mz2mass]{mz2mass()}} from the \code{MetaboCoreUtils} package.}

\item{ppm}{\code{numeric(1)} defining the allowed mass difference in ppm.}

\item{...}{additional parameters for the \code{compounds} call if \code{compounds} is
a \code{CompDb}.}

\item{mzcol}{for \code{object} being a \code{data.frame}: \code{character(1)} defining the
column name in \code{object} containing the m/z values. Defaults to \code{"mz"}.}
}
\value{
See description above for details.
}
\description{
\code{annotateMz} annotates m/z values using either a \code{data.frame} containing
compound definitions or a \code{\link[=CompDb]{CompDb()}} and a user provided list of potential
(expected) adducts. In detail, the function calculates for each input m/z
value the masses of the potential ion adducts and compares these with
(monoisotopic) masses provided in the reference annotation. A hit is returned
if the mass difference is smaller than \code{ppm}.

For \code{object} being a \code{numeric} and \code{compounds} a \code{data.frame} or equivalend
class, the function returns a \code{list} of subsets of \code{compounds} with rows
matching the adducts' mass. If no match is found for the mass of any
adducts, a 0-rows \code{data.frame} (same columns than \code{compounds} is returned).
}
\examples{

## Read compound annotations from a package-internal SDF file containing
## a small subset of HMDB annotations
fl <- system.file("sdf/HMDB_sub.sdf.gz", package = "CompoundDb")
cmps <- compound_tbl_sdf(fl)
cmps

## Annotate provided m/z values assuming they represent mass-to-charge
## values for [M+H]+ adducts.
mzs <- c(105.0546, 75.09168, 127.0365, 113.04756, 113.1210)

res <- annotateMz(mzs, cmps, adduct = "[M+H]+")

## res is a `list` with the matching hits for each m/z and the specified
## adduct. If no match was found the result contains only `NA` values.
res

## Using all adducts supported by the `mz2mass` function.
annotateMz(mzs, cmps, ppm = 10, adduct = MetaboCoreUtils::adductNames())

## It is also possible to annotate an input data.frame, in which case each
## input will be duplicated depending on the number of hits for each m/z
mzs_df <- data.frame(id = letters[1:5], mzmed = mzs)

res <- annotateMz(mzs_df, cmps, mzcol = "mzmed", adduct = "[M+H]+")
res

## Note that `compounds` could also be `CompDb` database, in which a search
## against the annotations provided in that database would be performed.
}
\author{
Johannes Rainer
}
