% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompDbPackage.R
\name{compound_tbl_sdf}
\alias{compound_tbl_sdf}
\title{Extract compound data from a file in SDF format}
\usage{
compound_tbl_sdf(file)
}
\arguments{
\item{file}{\code{character(1)} with the name of the SDF file.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with general compound information (one row per
compound):
\itemize{
\item \code{compound_id}: the ID of the compound.
\item \code{compound_name}: the compound's name.
\item \code{inchi}: the inchi of the compound.
\item \code{formula}: the chemical formula.
\item \code{mass}: the compound's mass.
}
}
\description{
\code{compound_tbl_sdf} extracts basic compound annotations from a file in SDF
format (structure-data file). The function currently supports SDF files from:
\itemize{
\item HMDB (Human Metabolome Database): http://www.hmdb.ca
\item ChEBI (Chemical Entities of Biological Interest): http://ebi.ac.uk/chebi
\item LMSD (LIPID MAPS Structure Database).
}
}
\examples{

## Read compound information from a subset of HMDB
fl <- system.file("sdf/HMDB_sub.sdf")
compound_tbl_sdf(fl)
}
\author{
Johannes Rainer and Jan Stanstrup
}
