% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-methods.R
\docType{methods}
\name{hasMz,Spectrum-method}
\alias{hasMz,Spectrum-method}
\alias{hasMz,Spectra-method}
\alias{hasMz}
\title{Determine whether a spectrum contains peaks with certain m/z}
\usage{
\S4method{hasMz}{Spectrum}(object, mz, ppm = 10, which = c("any",
  "all"))
}
\arguments{
\item{object}{\linkS4class{Spectrum} or \code{\link[=Spectra]{Spectra()}} object.}

\item{mz}{\code{numeric} with the query m/z value(s) to be looked for in \code{object}.}

\item{ppm}{\code{numeric(1)} defining the allowed deviation of the peak's m/z
from \code{mz}. By default (\code{ppm = 10}), a difference of +/-
\code{mz * 10/2e6} is allowed.}

\item{which}{\code{character(1)} defining whether peaks with m/z matching any
(\code{which = "any"}, default) of the provided \code{mz} are present, or all
(\code{which = "all"}). See examples for details.}
}
\value{
\code{logical} of length equal \code{x} whether a peak with the given m/z is
present in the spectrum.
}
\description{
\code{hasMz} allows to determine whether a spectrum contains peaks with their mz
matching any or all submitted \code{mz} values, accepting a certain deviation.
}
\examples{

sp1 <- new("Spectrum1", mz = c(23.231, 123.43, 255.231, 432.0952),
    intensity = c(123, 3432, 45432, 423))
sp2 <- new("Spectrum1", mz = c(123.099, 344.531, 453.2313),
    intensity = c(231, 431, 413))
sp3 <- new("Spectrum1", mz = c(123.1001, 343.4321, 432.0921),
    intensity = c(542, 4524, 32))

## Test for a single Spectrum if it contains a peak with an m/z matching
## any of the provided m/z.
mzs <- c(123.1, 432.0931)

## Does it contain any peak?
hasMz(sp1, mzs)

## Does it contain peaks matching all m/z?
hasMz(sp1, mzs, which = "all")

## Same for a list of Spectrum objects
spl <- Spectra(sp1, sp2, sp3)

## Do spectra contain any of the
hasMz(spl, mzs)
}
\author{
Johannes Rainer
}
