% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-utility-functions.R
\name{mass2mz}
\alias{mass2mz}
\alias{mz2mass}
\alias{adducts}
\title{Conversion functions between mass and m/z}
\usage{
mass2mz(x, adduct = adducts())

mz2mass(x, adduct = adducts())

adducts(pattern, polarity, name, set, ...)
}
\arguments{
\item{x}{\code{numeric} with the masses (or m/z values) to be converted.}

\item{adduct}{either a \code{data.frame} with required columns \code{"name"}, \code{"nmol"}
(number of molecules), \code{"charge"} (the total charge of the molecule)
and \code{"massdiff"} (total mass difference), such as returned by the
\code{adducts} function, or a \code{character} with the names of the adducts
(see \code{adducts} for supported adducts).}

\item{pattern}{For \code{adducts}: optional \code{character(1)} specifying a pattern to
be used to retrieve selected adducts, e.g. containing a hydrogen.}

\item{polarity}{For \code{adducts}: optional \code{numeric(1)} to retrieve only adducts
with positive (\code{polarity = 1}) or negative (\code{polarity = -1}) polarity.}

\item{name}{For \code{adducts}: define the names of the adducts for which the
definition should be returned.}

\item{set}{For \code{adducts}: \code{character(1)} defining sets of adducts.}

\item{...}{For \code{adducts}: additional parameters to be passed to the \code{\link[=grep]{grep()}}
function.}
}
\description{
\code{mass2mz} and \code{mz2mass} allow to convert between (monoisotopic) mass and m/z
of specified ion adducts an \emph{vice versa}. The functions return a \code{list} with
length equal to \code{x}, each element being a \code{numeric} with the m/z or mass for
the specified adducts. See below for examples.

\code{adducts} retrieves the definitions for the supported ion adducts. The
function returns a \code{data.frame}.
}
\examples{

masses <- c(75.032028409, 105.042595, 162.115698458, 180.063385)
names(masses) <- c("a", "b", "c", "d")

## Calculate mz for adducts [M+H]+ and [M+Na]+
mzs <- mass2mz(masses, adduct = c("[M+H]+", "[M+Na]+"))
}
\author{
Johannes Rainer, Jan Stanstrup
}
