% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-utility-functions.R
\name{matchWithPpm}
\alias{matchWithPpm}
\title{Match values allowing a certain difference in ppm}
\usage{
matchWithPpm(x, y, ppm = 0)
}
\arguments{
\item{x}{\code{numeric} with input values to match in \code{y}. Supposed to be a
shorter \code{numeric} than \code{y}.}

\item{y}{\code{numeric} with values to match \code{x} against.}

\item{ppm}{\code{numeric(1)} defining the allowed difference between values to be
still considered a match. Differences smaller than +/- \code{ppm} are
accepted.}
}
\value{
\code{list} with \code{integer} representing the index in \code{y} where each
element in \code{x} matches (given the provided ppm).
}
\description{
Match numeric values in \code{x} against values in \code{y} and return indices of
\strong{all} matches of \code{x} in \code{y} allowing differences between the values
specified with parameter \code{ppm} (part per million).
}
\examples{

yvals <- abs(rnorm(1000))
yvals[123] <- yvals[2] - yvals[2] * 10 * 1e-6
yvals[124] <- yvals[2] + yvals[2] * 10 * 1e-6
yvals[125] <- yvals[2] + yvals[2] * 12 * 1e-6
xvals <- yvals[c(2, 3, 3, 20, 21, 20)]
xvals[2] <- xvals[2] + (10 * xvals[2] / 1e6)
xvals[3] <- xvals[3] - (10 * xvals[3] / 1e6)
xvals[6] <- xvals[6] + (12 * xvals[6] / 1e6)

## Perfect matches:
matchWithPpm(xvals, yvals)

## Match allowing +/- 10ppm difference
matchWithPpm(xvals, yvals, ppm = 10)

## Match allowing +/- 20ppm difference
matchWithPpm(xvals, yvals, ppm = 20)
}
\author{
Johannes Rainer
}
