% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectrum2List.R, R/Spectrum2List-methods.R
\docType{methods}
\name{Spectrum2List}
\alias{Spectrum2List}
\alias{Spectrum2List-class}
\alias{show,Spectrum2List-method}
\alias{$,Spectrum2List-method}
\alias{$<-,Spectrum2List-method}
\alias{mz,Spectrum2List-method}
\alias{intensity,Spectrum2List-method}
\alias{rtime,Spectrum2List-method}
\title{List of Spectrum 2 objects along with annotations}
\usage{
Spectrum2List(...)

\S4method{$}{Spectrum2List}(x, name)

\S4method{$}{Spectrum2List}(x, name) <- value

\S4method{mz}{Spectrum2List}(object)

\S4method{intensity}{Spectrum2List}(object)

\S4method{rtime}{Spectrum2List}(object)
}
\arguments{
\item{...}{\link{Spectrum2} object(s), a \code{list} of \link{Spectrum2} objects or a
\code{data.frame}. See below for details.}

\item{x}{`Spectrum2List`.}

\item{name}{for `$` and `$<-`: `character(1)` with the name of the
metadata column to access.}

\item{value}{for `$<-`: vector to replace/set the specified column in the
object's metadata.}

\item{object}{`Spectrum2List`.}
}
\description{
\code{Spectrum2List} objects allow to collect one or more \link{Spectrum2} object(s)
in a \code{list}-like structure with the additional possibility to add arbitrary
annotations to each individual \link{Spectrum2} object. These can be accessed/set
with the \link{mcols} method.
}
\details{
\code{Spectrum2List} inherits all methods from the \link{SimpleList} class of the
\code{S4Vectors} package. This includes \code{lapply} and other data manipulation
and subsetting operations.
}
\section{Constructor}{


New \link{Spectrum2List} can be created with the \code{Spectrum2List(...)} function
where \code{...} can either be a single \link{Spectrum2} object, a \code{list} of
\link{Spectrum2} objects or a \code{data.frame}. If a \code{data.frame} is provided each
row is expected to represent one peak and the following required columns
have to be present:
\itemize{
\item spectrum_id: (\code{character}) defining the
\item mz: (\code{numeric}) the m/z value of each peak.
\item intensity: (\code{numeric}) the peak's intensity.
}

Optional columns listed below are used to fill the indicated slots of the
\link{Spectrum2} object:
\itemize{
\item polarity: \code{polarity}.
\item ms_level: \code{msLevel}.
\item rt: \code{rt}
\item precursor_mz: \code{precursorMz}.
\item precursor_charge: \code{precursorCharge}.
\item precursor_intensity: \code{precursorIntensity}.
\item collision_energy: \code{collisionEnergy}.
\item acquisition_num: \code{acquisitionNum}.
\item scan_index: \code{scanIndex}.
\item from_file: \code{fromFile}.
\item precursor_scan_num: \code{precScanNum}.
}

Any additional columns are added to the object's metadata columns \code{mcols}.
}

\section{Accessing metadata columns}{


Metadata columns can be accessed with `$`, that allows also to replace
columns or add new columns. The [DataFrame] with all metadata columns can
be accessed with the `mcols` function.
}

\section{Accessing spectrum attributes}{


These methods allow to access the attributes and values of the individual
`Spectrum2` objects. For more details please refer to the corresoonding
methods in the `MSnbase` package

- `mz` return the m/z values of each spectrum as a `list` of `numeric`
  vectors.

- `intensity` return the intensity values of each spectrum as a `list` of
  `numeric` vectors.

- `rtime` return the retention time of each spectrum as a `numeric` vector
  with length equal to the length of `object`.
}

\examples{

## Create from Spectrum2 objects
sp1 <- new("Spectrum2", mz = c(1, 2, 4), intensity = c(4, 5, 2))
sp2 <- new("Spectrum2", mz = c(1, 2, 3, 4), intensity = c(5, 3, 2, 5))

spl <- Spectrum2List(sp1, sp2)

## Build from data.frame
df <- data.frame(spectrum_id = c("b", "b", "b", "b", "a", "a", "a"),
    mz = c(1, 2, 3, 4, 1, 2, 4), intensity = c(5, 3, 2, 5, 4, 5, 2),
    polarity = c(1, 1, 1, 1, 0, 0, 0), compound_id = rep("cp_1", 7),
    stringsAsFactors = FALSE)

spl <- Spectrum2List(df)
polarity(spl[[1]])

## Additional annotations are provided in mcols
mcols(spl)

## Load Spectrum information from xml files
dr <- system.file("xml", package = "CompoundDb")

spl <- Spectrum2List(msms_spectra_hmdb(dr))

spl

## Access individual metadata columns using $
spl$spectrum_id

## Add additional metadata columns
spl$new_col <- 1:2

spl

## Extract the mz values for the individual spectra
mz(spl)

## Extract the intensity values for the individual spectra
intensity(spl)

## Extract the retention time values for the individual spectra
rtime(spl)
}
\seealso{
\link{Spectrum2} in the \code{MSnbase} package.

\code{\link[=msms_spectra_hmdb]{msms_spectra_hmdb()}} for a function to import a \code{data.frame} from files
containing spectrum MS/MS information from HMDB (http://www.hmdb.ca) in
xml format.
}
\author{
Johannes Rainer
}
