% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompDbPackage.R
\name{compound_tbl_lipidblast}
\alias{compound_tbl_lipidblast}
\title{Extract compound data from LipidBlast}
\usage{
compound_tbl_lipidblast(file, collapse)
}
\arguments{
\item{file}{\code{character(1)} with the name of the file name.}

\item{collapse}{optional \code{character(1)} to be used to collapse multiple
values in the columns \code{"synonyms"}. See examples for details.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with general compound information (one row per
compound):
\itemize{
\item \code{compound_id}: the ID of the compound.
\item \code{name}: the compound's name.
\item \code{inchi}: the InChI of the compound.
\item \code{inchikey}: the InChI key.
\item \code{formula}: the chemical formula.
\item \code{exactmass}: the compound's mass.
\item \code{synonyms}: the compound's synonyms (aliases). This type of this column is
by default a \code{list} to support multiple aliases per compound, unless
argument \code{collapse} is provided, in which case multiple synonyms are pasted
into a single element separated by the value of \code{collapse}.
}
}
\description{
\code{compound_tbl_lipidblast()} extracts basic comopund annotations from a
LipidBlast file in (json format) downloaded from
http://mona.fiehnlab.ucdavis.edu/downloads
}
\examples{

## Read compound information from a subset of HMDB
fl <- system.file("json/MoNa-LipidBlast_sub.json", package = "CompoundDb")
cmps <- compound_tbl_lipidblast(fl)
cmps
}
\seealso{
Other compound table creation functions: 
\code{\link{compound_tbl_sdf}()}
}
\author{
Johannes Rainer and Jan Stanstrup
}
\concept{compound table creation functions}
