% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFilters.R
\docType{class}
\name{Filter-classes}
\alias{Filter-classes}
\alias{CompoundIdFilter-class}
\alias{CompoundIdFilter}
\alias{CompoundNameFilter-class}
\alias{CompoundNameFilter}
\alias{MsmsMzRangeMinFilter-class}
\alias{MsmsMzRangeMinFilter}
\alias{MsmsMzRangeMaxFilter-class}
\alias{MsmsMzRangeMaxFilter}
\title{Filters supported by CompDb}
\usage{
CompoundIdFilter(value, condition = "==")

CompoundNameFilter(value, condition = "==")

MsmsMzRangeMinFilter(value, condition = ">=")

MsmsMzRangeMaxFilter(value, condition = "<=")
}
\arguments{
\item{value}{The value for the filter. For details see
\code{\link[AnnotationFilter:AnnotationFilter]{AnnotationFilter::AnnotationFilter()}}.}

\item{condition}{The condition for the filter. For details see
\code{\link[AnnotationFilter:AnnotationFilter]{AnnotationFilter::AnnotationFilter()}}.}
}
\description{
A variety of different filters can be applied to the \code{CompDb} object to
retrieve only subsets of the data. These filters extend the
\link[AnnotationFilter:AnnotationFilter]{AnnotationFilter::AnnotationFilter} class and support the filtering concepts
introduced by Bioconductor's \code{AnnotationFilter} package.

The supported filters are:
\itemize{
\item \code{CompoundIdFilter}: filter based on the compound ID.
\item \code{CompoundNameFilter}: filter based on the compound name.
\item \code{MsmsMzRangeMinFilter}: retrieve entries based on the smallest m/z of all
peaks of their MS/MS spectra. Requires that MS/MS spectra data are present
(i.e. \code{hasMsMsSpectra(cmp_db)} returns \code{TRUE}).
\item \code{MsmsMzRangeMaxFilter}: retrieve entries based on the largest m/z of all
peaks of their MS/MS spectra. Requires that MS/MS spectra data are present
(i.e. \code{hasMsMsSpectra(cmp_db)} returns \code{TRUE}).
}
}
\examples{
library(CompoundDb)

## Create a filter for the compound id
cf <- CompoundIdFilter("comp_a")
cf

## Create a filter using a formula expression
AnnotationFilter(~ compound_id == "comp_b")

## Combine filters
AnnotationFilterList(CompoundIdFilter("a"), CompoundNameFilter("b"))

## Using a formula expression
AnnotationFilter(~ compound_id == "a" | compound_name != "b")
}
\seealso{
\code{\link[=supportedFilters]{supportedFilters()}} for the method to list all supported filters
for a \code{CompDb} object.
}
\author{
Johannes Rainer
}
