% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendCompDb-functions.R,
%   R/MsBackendCompDb.R
\docType{class}
\name{MsBackendCompDb}
\alias{MsBackendCompDb}
\alias{MsBackendCompDb-class}
\alias{backendInitialize,MsBackendCompDb-method}
\alias{show,MsBackendCompDb-method}
\alias{peaksData,MsBackendCompDb-method}
\alias{intensity,MsBackendCompDb-method}
\alias{intensity<-,MsBackendCompDb-method}
\alias{mz,MsBackendCompDb-method}
\alias{mz<-,MsBackendCompDb-method}
\alias{spectraData,MsBackendCompDb-method}
\alias{spectraData<-,MsBackendCompDb-method}
\alias{$<-,MsBackendCompDb-method}
\title{CompDb-based MS spectrum backend}
\usage{
MsBackendCompDb()

\S4method{backendInitialize}{MsBackendCompDb}(object, x, columns, filter, ...)

\S4method{show}{MsBackendCompDb}(object)

\S4method{peaksData}{MsBackendCompDb}(object)

\S4method{intensity}{MsBackendCompDb}(object)

\S4method{intensity}{MsBackendCompDb}(object) <- value

\S4method{mz}{MsBackendCompDb}(object)

\S4method{mz}{MsBackendCompDb}(object) <- value

\S4method{spectraData}{MsBackendCompDb}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendCompDb}(object) <- value

\S4method{$}{MsBackendCompDb}(x, name) <- value
}
\arguments{
\item{object}{an \code{MsBackendCompDb} instance.}

\item{x}{an \code{MsBackendCompDb} instance.}

\item{columns}{for \code{spectraData}: \code{character} with names of columns/spectra
variables that should be returned. Defaults to
\code{spectraVariables(object)}. Database columns \code{"ms_level"},
\code{"precursor_mz"}, \code{"precursor_intensity"}, \code{"precursor_charge"} are
mapped to the core \code{Spectra} variables \code{msLevel}, \code{precursorMz},
\code{precursorIntensity} and \code{precursorCharge}, respectively.}

\item{filter}{for \code{backendInitialize}: optional filter expression to specify
which elements to retrieve from the database.}

\item{...}{ignored.}

\item{value}{for \verb{$<-}: the replacement values.}

\item{name}{for \verb{$<-}: the name of the spectra variable to replace.}
}
\value{
See the description of the respective function.
}
\description{
The \code{MsBackendCompDb} allows to retrieve MS2 spectra from an \code{\link[=CompDb]{CompDb()}}
object/database. The object keeps only a limited amount of data in memory
and retrieves the m/z and intensity values from the database \emph{on-demand}.

It is not intended that users create or use instances of this class directly,
the \code{\link[=Spectra]{Spectra()}} call on \code{\link[=CompDb]{CompDb()}} will return a \code{Spectra} object that uses
this backend.
}
\note{
For higher performance it is suggested to change the backend of the
\code{\link[=Spectra]{Spectra()}} object to an \code{\link[=MsBackendDataFrame]{MsBackendDataFrame()}} backend with the
\code{\link[=setBackend]{setBackend()}} method of \code{Spectra} objects.
}
\section{Methods implemented for \code{MsBackendCompDb}}{


The methods listed here are implemented for the \code{MsBackendCompDb}. All other
methods are inherited directly from the parent \code{\link[=MsBackendDataFrame]{MsBackendDataFrame()}} class.
See the help of \code{MsBackendDataFrame} in the \code{Spectra} package for a
complete listing of methods.
\itemize{
\item \code{peaksData}: gets the full list of peak matrices. Returns a \code{\link[=list]{list()}},
length equal to the number of spectra and each element being a \code{matrix}
with columns \code{"mz"} and \code{"intensity"} with the spectra's m/z and intensity
values.
\item \code{intensity}: retrieves the intensity values for all spectra. Returns a
\code{\link[=NumericList]{NumericList()}}, each element being the intensity values of one spectrum.
The actual intensity values are retrieved from the database.
\item \verb{intensity<-}: not supported.
\item \code{mz}: retrieves the m/z values for all spectra. Returns a \code{\link[=NumericList]{NumericList()}},
each element being a \code{numeric} with the m/z values of one spectrum. These
values are retrieved from the database.
\item \verb{mz<-}: not supported.
\item \code{spectraData}: returns the complete spectrum data including m/z and
intensity values as a \code{\link[=DataFrame]{DataFrame()}}.
\item \verb{spectraData<-}: replace the spectrum metadata. Note that columns \code{"mz"}
and \code{"intensity"} are ignored.
\item \verb{$<-}: replace or add a spectrum variable. Note that \code{mz}, \code{intensity} and
\code{spectrum_id} variables can not be replaced.
}
}

\examples{

## MsBackendCompDb are not expected to be created/instanciated by users
## directly. Users also almost never directly interact with this type of
## object, as it is intented as a pure data backend for the `Spectra` object.
## Users will thus access MS data through such `Spectra` object, which can
## be created for `CompDb` objects using the `Spectra` method (see help
## of the `CompDb` object for more information. This examples shows how
## a `MsBackendCompDb` could be created purely from an SQLite database
## with data from a CompoundDb database.

## Connect to the SQLite database of a `CompDb` distributed via this package
library(RSQLite)
library(Spectra)
cdb <- CompDb(system.file("sql/CompDb.MassBank.sql", package = "CompoundDb"))

be <- backendInitialize(MsBackendCompDb(), cdb)
be

## Accessing m/z values
mz(be)
}
\author{
Johannes Rainer
}
