% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendCompDb-functions.R,
%   R/MsBackendCompDb.R
\docType{class}
\name{MsBackendCompDb}
\alias{MsBackendCompDb}
\alias{MsBackendCompDb-class}
\alias{backendInitialize,MsBackendCompDb-method}
\alias{show,MsBackendCompDb-method}
\alias{as.list,MsBackendCompDb-method}
\alias{intensity,MsBackendCompDb-method}
\alias{intensity<-,MsBackendCompDb-method}
\alias{mz,MsBackendCompDb-method}
\alias{mz<-,MsBackendCompDb-method}
\alias{asDataFrame,MsBackendCompDb-method}
\alias{asDataFrame<-,MsBackendCompDb-method}
\alias{$<-,MsBackendCompDb-method}
\title{CompDb-based MS spectrum backend}
\usage{
MsBackendCompDb()

\S4method{backendInitialize}{MsBackendCompDb}(object, x, columns, filter, ...)

\S4method{show}{MsBackendCompDb}(object)

\S4method{as.list}{MsBackendCompDb}(x)

\S4method{intensity}{MsBackendCompDb}(object)

\S4method{intensity}{MsBackendCompDb}(object) <- value

\S4method{mz}{MsBackendCompDb}(object)

\S4method{mz}{MsBackendCompDb}(object) <- value

\S4method{asDataFrame}{MsBackendCompDb}(object, columns = spectraVariables(object))

\S4method{asDataFrame}{MsBackendCompDb}(object) <- value

\S4method{$}{MsBackendCompDb}(x, name) <- value
}
\arguments{
\item{object}{an \code{MsBackendCompDb} instance.}

\item{x}{an \code{MsBackendCompDb} instance.}

\item{columns}{for \code{asDataFrame}: \code{character} with names of columns/spectra
variables that should be returned. Defaults to
\code{spectraVariables(object)}.}

\item{filter}{for \code{backendInitialize}: optional filter expression to specify
which elements to retrieve from the database.}

\item{...}{ignored.}

\item{value}{for \verb{$<-}: the replacement values.}

\item{name}{for \verb{$<-}: the name of the spectra variable to replace.}
}
\description{
The \code{MsBackendCompDb} allows to retrieve MS2 spectra from an \code{\link[=CompDb]{CompDb()}}
object/database. The object keeps only a limited amount of data in memory
and retrieves the m/z and intensity values from the database \emph{on-demand}.

It is not intended that users create or use instances of this class directly,
the \code{\link[=Spectra]{Spectra()}} call on \code{\link[=CompDb]{CompDb()}} will return a \code{Spectra} object that uses
this backend.
}
\note{
For higher performance it is suggested to change the backend of the
\code{\link[=Spectra]{Spectra()}} object to an \code{\link[=MsBackendDataFrame]{MsBackendDataFrame()}} backend with the
\code{\link[=setBackend]{setBackend()}} method of \code{Spectra} objects.
}
\section{Methods implemented for \code{MsBackendCompDb}}{


The methods listed here are implemented for the \code{MsBackendCompDb}. All other
methods are inherited directly from the parent \code{\link[=MsBackendDataFrame]{MsBackendDataFrame()}} class.
See that help page for a complete listing of methods.
\itemize{
\item \code{as.list}: gets the full list of peak matrices. Returns a \code{\link[=SimpleList]{SimpleList()}},
length equal to the number of spectra and each element being a \code{matrix}
with columns \code{"mz"} and \code{"intensity"} with the spectra's m/z and intensity
values.
\item \code{intensity}: retrieves the intensity values for all spectra. Returns a
\code{\link[=NumericList]{NumericList()}}, each element being the intensity values of one spectrum.
The actual intensity values are retrieved from the database.
\item \verb{intensity<-}: not supported.
\item \code{mz}: retrieves the m/z values for all spectra. Returns a \code{\link[=NumericList]{NumericList()}},
each element being a \code{numeric} with the m/z values of one spectrum. These
values are retrieved from the database.
\item \verb{mz<-}: not supported.
\item \code{asDataFrame}: returns the complete spectrum data including m/z and
intensity values as a \code{\link[=DataFrame]{DataFrame()}}.
\item \verb{asDataFrame<-}: replace the spectrum metadata. Note that columns \code{"mz"}
and \code{"intensity"} are ignored.
\item \verb{$<-}: replace or add a spectrum variable. Note that \code{mz}, \code{intensity} and
\code{spectrum_id} variables can not be replaced.
}
}

\author{
Johannes Rainer
}
