% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectrum2List.R
\name{Spectrum2List}
\alias{Spectrum2List}
\alias{Spectrum2List-class}
\alias{show,Spectrum2List-method}
\alias{Spectrum2List}
\title{List of Spectrum 2 objects along with annotations}
\usage{
Spectrum2List(...)
}
\arguments{
\item{...}{\link{Spectrum2} object(s), a \code{list} of \link{Spectrum2} objects or a
\code{data.frame}. See descritpion for details.}
}
\description{
\code{Spectrum2List} objects allow to collect one or more \link{Spectrum2} object(s)
in a \code{list}-like structure with the additional possibility to add arbitrary
annotations to each individual \link{Spectrum2} object. These can be accessed/set
with the \link{mcols} method.
}
\section{Constructor}{


New \link{Spectrum2List} can be created with the \code{Spectrum2List(...)} function
where \code{...} can either be a single \link{Spectrum2} object, a \code{list} of
\link{Spectrum2} objects or a \code{data.frame}. If a \code{data.frame} is provided each
row is expected to represent one peak and the following required columns
have to be present:
\itemize{
\item spectrum_id: (\code{character}) defining the
\item mz: (\code{numeric}) the m/z value of each peak.
\item intensity: (\code{numeric}) the peak's intensity.
}

Optional columns listed below are used to fill the indicated slots of the
\link{Spectrum2} object:
\itemize{
\item polarity: \code{polarity}.
\item ms_level: \code{msLevel}.
\item rt: \code{rt}
\item precursor_mz: \code{precursorMz}.
\item precursor_charge: \code{precursorCharge}.
\item precursor_intensity: \code{precursorIntensity}.
\item collision_energy: \code{collisionEnergy}.
\item acquisition_num: \code{acquisitionNum}.
\item scan_index: \code{scanIndex}.
\item from_file: \code{fromFile}.
\item precursor_scan_num: \code{precScanNum}.
}

Any additional columns are added to the object's metadata columns \code{mcols}.
}

\examples{

## Create from Spectrum2 objects
sp1 <- new("Spectrum2", mz = c(1, 2, 4), intensity = c(4, 5, 2))
sp2 <- new("Spectrum2", mz = c(1, 2, 3, 4), intensity = c(5, 3, 2, 5))

spl <- Spectrum2List(sp1, sp2)

## Build from data.frame
df <- data.frame(spectrum_id = c("b", "b", "b", "b", "a", "a", "a"),
    mz = c(1, 2, 3, 4, 1, 2, 4), intensity = c(5, 3, 2, 5, 4, 5, 2),
    polarity = c(1, 1, 1, 1, 0, 0, 0), compound_id = rep("cp_1", 7),
    stringsAsFactors = FALSE)

spl <- Spectrum2List(df)
polarity(spl[[1]])

## Additional annotations are provided in mcols
mcols(spl)
}
\seealso{
\link{Spectrum2} in the \code{MSnbase} package.
}
\author{
Johannes Rainer

Johannes Rainer
}
