% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectrum2List.R, R/Spectrum2List-methods.R
\docType{methods}
\name{Spectrum2List}
\alias{Spectrum2List}
\alias{Spectrum2List-class}
\alias{show,Spectrum2List-method}
\alias{$,Spectrum2List-method}
\alias{$<-,Spectrum2List-method}
\alias{mz,Spectrum2List-method}
\alias{intensity,Spectrum2List-method}
\alias{rtime,Spectrum2List-method}
\alias{precursorMz,Spectrum2List-method}
\alias{precursorCharge,Spectrum2List-method}
\alias{precScanNum,Spectrum2List-method}
\alias{precursorIntensity,Spectrum2List-method}
\alias{acquisitionNum,Spectrum2List-method}
\alias{scanIndex,Spectrum2List-method}
\alias{peaksCount,Spectrum2List,ANY-method}
\alias{msLevel,Spectrum2List-method}
\alias{tic,Spectrum2List-method}
\alias{ionCount,Spectrum2List-method}
\alias{collisionEnergy,Spectrum2List-method}
\alias{fromFile,Spectrum2List-method}
\alias{polarity,Spectrum2List-method}
\alias{smoothed,Spectrum2List-method}
\alias{isEmpty,Spectrum2List-method}
\alias{centroided,Spectrum2List-method}
\alias{isCentroided,Spectrum2List-method}
\title{List of Spectrum 2 objects along with annotations}
\usage{
Spectrum2List(...)

\S4method{$}{Spectrum2List}(x, name)

\S4method{$}{Spectrum2List}(x, name) <- value

\S4method{mz}{Spectrum2List}(object)

\S4method{intensity}{Spectrum2List}(object)

\S4method{rtime}{Spectrum2List}(object)

\S4method{precursorMz}{Spectrum2List}(object)

\S4method{precursorCharge}{Spectrum2List}(object)

\S4method{precScanNum}{Spectrum2List}(object)

\S4method{precursorIntensity}{Spectrum2List}(object)

\S4method{acquisitionNum}{Spectrum2List}(object)

\S4method{scanIndex}{Spectrum2List}(object)

\S4method{peaksCount}{Spectrum2List,ANY}(object)

\S4method{msLevel}{Spectrum2List}(object)

\S4method{tic}{Spectrum2List}(object)

\S4method{ionCount}{Spectrum2List}(object)

\S4method{collisionEnergy}{Spectrum2List}(object)

\S4method{fromFile}{Spectrum2List}(object)

\S4method{polarity}{Spectrum2List}(object)

\S4method{smoothed}{Spectrum2List}(object)

\S4method{isEmpty}{Spectrum2List}(x)

\S4method{centroided}{Spectrum2List}(object)

\S4method{isCentroided}{Spectrum2List}(object)
}
\arguments{
\item{...}{\link{Spectrum2} object(s), a \code{list} of \link{Spectrum2} objects or a
\code{data.frame}. See below for details.}

\item{x}{`Spectrum2List`.}

\item{name}{for `$` and `$<-`: `character(1)` with the name of the
metadata column to access.}

\item{value}{for `$<-`: vector to replace/set the specified column in the
object's metadata.}

\item{object}{`Spectrum2List`.}
}
\description{
\code{Spectrum2List} objects allow to collect one or more \link{Spectrum2} object(s)
in a \code{list}-like structure with the additional possibility to add arbitrary
annotations to each individual \link{Spectrum2} object. These can be accessed/set
with the \link{mcols} method.
}
\details{
\code{Spectrum2List} inherits all methods from the \link{SimpleList} class of the
\code{S4Vectors} package. This includes \code{lapply} and other data manipulation
and subsetting operations.
}
\section{Constructor}{


New \link{Spectrum2List} can be created with the \code{Spectrum2List(...)} function
where \code{...} can either be a single \link{Spectrum2} object, a \code{list} of
\link{Spectrum2} objects or a \code{data.frame}. If a \code{data.frame} is provided each
row is expected to represent one peak and the following required columns
have to be present:
\itemize{
\item spectrum_id: (\code{character}) defining the
\item mz: (\code{numeric} or \code{list} of \code{numeric}) the m/z value of each peak.
\item intensity: (\code{numeric} or \code{list} of \code{numeric}) the peak's intensity.
}

Optional columns listed below are used to fill the indicated slots of the
\link{Spectrum2} object:
\itemize{
\item polarity: \code{polarity}.
\item ms_level: \code{msLevel}.
\item rt: \code{rt}
\item precursor_mz: \code{precursorMz}.
\item precursor_charge: \code{precursorCharge}.
\item precursor_intensity: \code{precursorIntensity}.
\item collision_energy: \code{collisionEnergy}.
\item acquisition_num: \code{acquisitionNum}.
\item scan_index: \code{scanIndex}.
\item from_file: \code{fromFile}.
\item precursor_scan_num: \code{precScanNum}.
}

Any additional columns are added to the object's metadata columns \code{mcols}.
}

\section{Accessing metadata columns}{


Metadata columns can be accessed with `$`, that allows also to replace
columns or add new columns. The [DataFrame] with all metadata columns can
be accessed with the `mcols` function.
}

\section{Accessing spectrum attributes}{


These methods allow to access the attributes and values of the individual
`Spectrum2` objects. For more details please refer to the corresponding
methods in the `MSnbase` package

- `mz` return the m/z values of each spectrum as a `list` of `numeric`
  vectors.

- `intensity` return the intensity values of each spectrum as a `list` of
  `numeric` vectors.

- `rtime` return the retention time of each spectrum as a `numeric` vector
  with length equal to the length of `object`.

- `precursorMz`, `precursorCharge`, `precursorIntensity`, `precScanNum`
  return precursor m/z values, charge, intensity and scan number for each
  spectrum as a `numeric` (or `integer`) vector with length equal to the
  length of `object`.

- `acquisitionNum` and `scanIndex` return the acquisition number of each
  spectrum and its scan index as an `integer` vector with the same length
  than `object`.

- `ionCount` and `tic` return the ion count and total ion current of each
  spectrum.

- `peaksCount` returns the number of peaks for each spectrum as a `integer`
  vector.

- `msLevel` returns the MS level of each spectrum.

- `collisionEnergy` returns the collision energy for each spectrum.

- `polarity` returns the spectra's polarity.

- `fromFile` returns the index from the (e.g. mzML) file the spectra where
  from. This applies only for spectra read using e.g. `MSnbase`'s
  `readMSData` function.

- `smoothed` whether spectra have been smoothed (i.e. processed with the
  [MSnbase::smooth] method. Returns a `logical` of length equal to the
  number of spectra.

- `isEmpty` returns `TRUE` for spectra without peak data.

- `centroided`, `isCentroided` returns for each spectrum whether it contains
  *centroided* data. While `centroided` returns the internal attribute of
  each spectrum, `isCentroided` tries to guess whether spectra are
  centroided from the actual peak data.
}

\examples{

## Create from Spectrum2 objects
sp1 <- new("Spectrum2", mz = c(1, 2, 4), intensity = c(4, 5, 2))
sp2 <- new("Spectrum2", mz = c(1, 2, 3, 4), intensity = c(5, 3, 2, 5))

spl <- Spectrum2List(sp1, sp2)

## Build from data.frame
df <- data.frame(spectrum_id = c("b", "b", "b", "b", "a", "a", "a"),
    mz = c(1, 2, 3, 4, 1, 2, 4), intensity = c(5, 3, 2, 5, 4, 5, 2),
    polarity = c(1, 1, 1, 1, 0, 0, 0), compound_id = rep("cp_1", 7),
    stringsAsFactors = FALSE)

spl <- Spectrum2List(df)
polarity(spl[[1]])

## Additional annotations are provided in mcols
mcols(spl)

## Load Spectrum information from xml files
dr <- system.file("xml", package = "CompoundDb")

spl <- Spectrum2List(msms_spectra_hmdb(dr))

spl

## Access individual metadata columns using $
spl$spectrum_id

## Add additional metadata columns
spl$new_col <- 1:2

spl

## Extract the mz values for the individual spectra
mz(spl)

## Extract the intensity values for the individual spectra
intensity(spl)

## Extract the retention time values for the individual spectra
rtime(spl)

## Extract the precursor m/z of each spectrum.
precursorMz(spl)

## Extract the precursor charge of each spectrum.
precursorCharge(spl)

## Extract the precursor scan number for each spectrum.
precScanNum(spl)

## Extract the precursor intensity of each spectrum.
precursorIntensity(spl)

## Extract the acquisition number of each spectrum.
acquisitionNum(spl)

## Extract the scan index of each spectrum.
scanIndex(spl)

## Get the number of peaks per spectrum.
peaksCount(spl)

## Get the MS level of each spectrum.
msLevel(spl)

## Get the total ion current for each spectrum.
tic(spl)

## Get the total ion current for each spectrum.
ionCount(spl)

## Extract the collision energy for spectrum.
collisionEnergy(spl)

## Extract the file index for the spectrum.
fromFile(spl)

## Get the polarity for each spectrum.
polarity(spl)

## Whether spectra are smoothed (i.e. processed with the `MSnbase::smooth`
## function.
smoothed(spl)

## Are spectra empty (i.e. contain no peak data)?
isEmpty(spl)

## Do the spectra contain centroided data?
centroided(spl)

## Do the spectra contain centroided data? Whether spectra are centroided
## is estimated from the peak data.
isCentroided(spl)
}
\seealso{
\link{Spectrum2} in the \code{MSnbase} package.

\code{\link[=msms_spectra_hmdb]{msms_spectra_hmdb()}} for a function to import a \code{data.frame} from files
containing spectrum MS/MS information from HMDB (http://www.hmdb.ca) in
xml format.
}
\author{
Johannes Rainer
}
